chore: pages 폴더에 있는 모든 파일 components 폴더로 이관 #188
