+++ src/components/recruit/write/RecruitWriteSkeleton.tsx
@@ -1,0 +1,5 @@
+const RecruitWriteSkeleton = () => {
+  return <div></div>
+}
+
+export default RecruitWriteSkeleton
