+++ src/components/recruit/write/_dummyGetStudyGroupsResponse.ts
